+++ lib/src/console/logarte_dashboard_screen.dart
@@ -1,14 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:logarte/logarte.dart';
 import 'package:logarte/src/console/logarte_entry_item.dart';
+import 'package:logarte/src/console/logarte_fab_state.dart';
 import 'package:logarte/src/console/logarte_theme_wrapper.dart';
 
 class LogarteDashboardScreen extends StatefulWidget {
   final Logarte instance;
   final bool showBackButton;
+
   const LogarteDashboardScreen(
     this.instance, {
-    Key? key, this.showBackButton = false
+    Key? key,
+    this.showBackButton = false,
   }) : super(key: key);
 
   @override
@@ -22,11 +25,13 @@ class _LogarteDashboardScreenState extends State<LogarteDashboardScreen> {
   void initState() {
     super.initState();
     _controller = TextEditingController();
+    LogarteFabState.instance.open();
   }
 
   @override
   void dispose() {
     _controller.dispose();
+    LogarteFabState.instance.close();
     super.dispose();
   }
 
@@ -34,7 +39,7 @@ class _LogarteDashboardScreenState extends State<LogarteDashboardScreen> {
   Widget build(BuildContext context) {
     return LogarteThemeWrapper(
       child: DefaultTabController(
-        length: 5,
+        length: widget.instance.customTab != null ? 6 : 5,
         child: Scaffold(
           body: NestedScrollView(
             headerSliverBuilder: (context, innerBoxIsScrolled) {
@@ -86,6 +91,11 @@ class _LogarteDashboardScreenState extends State<LogarteDashboardScreen> {
                         text:
                             'Navigation (${widget.instance.logs.value.whereType<NavigatorLogarteEntry>().length})',
                       ),
+                      if (widget.instance.customTab != null)
+                        const Tab(
+                          icon: Icon(Icons.extension_rounded),
+                          text: 'Custom',
+                        ),
                     ],
                   ),
                 ),
@@ -122,6 +132,8 @@ class _LogarteDashboardScreenState extends State<LogarteDashboardScreen> {
                           instance: widget.instance,
                           search: search,
                         ),
+                        if (widget.instance.customTab != null)
+                          widget.instance.customTab!,
                       ],
                     );
                   },