+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.0.0
+
+* Hitting first stable release! 🎉 Logarte is now "readier" for production use.
+* Fixed a bug where clicking floating button in network details page would stuck the console.
+
+## 0.3.1
+
+* Added `disableDebugConsoleLogs` parameter to disable logs in IDE's debug console (default is false).
+
+## 0.3.0
+
+* You can now pass a custom tab to the console.
+
 ## 0.2.4
 
 * Improve database log's format and fix overflow issue on console.